Fix improper handling of null terminated strings
C like strings end with the first null byte, after that garbage could be the in the stream.
Currently only the streaming API did a TrimEnd for null bytes but even that could fail when the garbage appeared after the null byte.

This commit changes both the streaming api and the non streaming api to look for the first null byte and cut the string off after that.
If no null byte was found the whole input will be used, which should technically never happen as the client always makes spaces for the null byte.